forms refactor - set together section by section
