+++ libs/application/templates/estate/src/fields/DoneImage/assets/Grieving.tsx
