--- libs/application/templates/estate/src/forms/PrivateEstateDivision/done.ts
@@ -1,28 +1,0 @@
-import {
-  buildForm,
-  buildCustomField,
-  buildMultiField,
-} from '@island.is/application/core'
-import { Form, FormModes } from '@island.is/application/types'
-import { m } from '../../lib/messages'
-
-export const done: Form = buildForm({
-  id: 'divisionOfEstateByHeirs',
-  title: '',
-  mode: FormModes.COMPLETED,
-  renderLastScreenButton: true,
-  children: [
-    buildMultiField({
-      id: 'done',
-      title: m.doneTitle,
-      description: m.divisionOfEstateByHeirsSubtitle,
-      children: [
-        buildCustomField({
-          id: 'doneImage',
-          component: 'DoneImage',
-          title: '',
-        }),
-      ],
-    }),
-  ],
-})
