+++ libs/application/templates/estate/src/forms/Sections/approveSubmission.ts
@@ -1,0 +1,40 @@
+import {
+  buildDescriptionField,
+  buildMultiField,
+  buildCheckboxField,
+  buildSection,
+} from '@island.is/application/core'
+import { m } from '../../lib/messages'
+import { YES } from '../../lib/constants'
+
+export const approvePrivateDivisionSubmission = buildSection({
+  id: 'approveSubmission',
+  title: m.divisionOfEstateByHeirsTerms,
+  children: [
+    buildMultiField({
+      id: 'approveSubmission',
+      title: m.divisionOfEstateByHeirsTerms,
+      description: m.divisionOfEstateByHeirsText,
+      children: [
+        buildDescriptionField({
+          id: 'space',
+          title: '',
+          space: 'containerGutter',
+        }),
+        buildCheckboxField({
+          id: 'confirmAction',
+          title: '',
+          large: false,
+          backgroundColor: 'white',
+          defaultValue: [],
+          options: [
+            {
+              value: YES,
+              label: m.divisionOfEstateByHeirsSubmissionCheckbox.defaultMessage,
+            },
+          ],
+        }),
+      ],
+    }),
+  ],
+})
