+++ libs/application/templates/estate/src/forms/Sections/estateDebts.ts
@@ -1,0 +1,56 @@
+import {
+  buildSection,
+  buildMultiField,
+  buildCustomField,
+} from '@island.is/application/core'
+import { m } from '../../lib/messages'
+import { EstateTypes, YES } from '../../lib/constants'
+
+export const estateDebts = buildSection({
+  id: 'debts',
+  title: m.debtsTitle,
+  condition: (answers) =>
+    answers.selectedEstate === EstateTypes.estateWithoutAssets
+      ? answers.estateAssetsExist === YES
+      : true,
+  children: [
+    buildMultiField({
+      id: 'debts',
+      title: m.debtsTitle,
+      description: m.debtsDescription,
+      children: [
+        buildCustomField(
+          {
+            title: '',
+            id: 'debts',
+            component: 'TextFieldsRepeater',
+          },
+          {
+            fields: [
+              {
+                title: m.debtsCreditorName.defaultMessage,
+                id: 'creditorName',
+              },
+              {
+                title: m.debtsNationalId.defaultMessage,
+                id: 'nationalId',
+                format: '######-####',
+              },
+              {
+                title: m.debtsLoanIdentity.defaultMessage,
+                id: 'loanIdentity',
+              },
+              {
+                title: m.debtsBalance.defaultMessage,
+                id: 'balance',
+                currency: true,
+              },
+            ],
+            repeaterButtonText: m.debtsRepeaterButton.defaultMessage,
+            repeaterHeaderText: m.debtsCreditorHeader.defaultMessage,
+          },
+        ),
+      ],
+    }),
+  ],
+})
