+++ libs/application/templates/estate/src/forms/Sections/estateMembers.ts
@@ -1,0 +1,31 @@
+import {
+  buildDescriptionField,
+  buildSection,
+  buildMultiField,
+  buildCustomField,
+} from '@island.is/application/core'
+import { m } from '../../lib/messages'
+
+export const estateMembers = buildSection({
+  id: 'estateMembersInfo',
+  title: m.estateMembers,
+  children: [
+    buildMultiField({
+      id: 'estateMembersInfo',
+      title: m.estateMembers,
+      description: m.estateMembersSubtitle,
+      children: [
+        buildCustomField({
+          title: '',
+          id: 'estate.estateMembers',
+          component: 'EstateMembersRepeater',
+        }),
+        buildDescriptionField({
+          id: 'space2',
+          title: '',
+          space: 'containerGutter',
+        }),
+      ],
+    }),
+  ],
+})
